moved HomeScreen to src/screens
