+++ App.js
@@ -1,7 +1,6 @@
-import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import { View, StyleSheet, Text, Button } from 'react-native';
+import HomeScreen from './src/screens/HomeScreen';
 import CreateHousholdScreen from './src/screens/CreateHouseholdScreen';
 
 const Stack = createStackNavigator();
@@ -16,41 +15,3 @@ export default function App() {
     </NavigationContainer>
   );
 }
-
-function HomeScreen({ navigation }) {
-  return (
-    <View style={styles.container}>
-      <Text>Welcome to the Shared Shopping App!</Text>
-      {/* CAN ADJUST LOCATION OF BUTTON IF NEEDED LATER */}
-      <View style={styles.buttonContainer}>
-        <Button
-          title="Create Household"
-          onPress={() => navigation.navigate('Create Household')}  
-        />
-      </View>
-    </View>
-  );
-}
-
-// title and buttonContainer used for adjusting BUTTON LOCATION
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: '#fffff',  // Changing the background to red to test
-  },
-  title: {
-    fontSize: 24,
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  buttonContainer: {
-    marginTop: 50,
-    alignSelf: 'center',
-    paddingHorizontal: 20,
-    position: 'absolute',
-    bottom: 325,
-  },
-});
-
